move assert_entity_golden helper
